src/main/java/mahasiswa/mahasiswa.java
/*
 * To change this license header, choose License Headers in Project Properties.
 * To change this template file, choose Tools | Templates
 * and open the template in the editor.
 */
package mahasiswa;

/**
 *
 * @author SB604-19
 */
public class mahasiswa {
    private String nrp;
    private String nama;
    private int umur;

    mahasiswa() {
        throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
    }

    public String getNrp() {
        return nrp;
    }

    public void setNrp(String nrp) {
        this.nrp = nrp;
    }

    public String getNama() {
        return nama;
    }

    public void setNama(String nama) {
        this.nama = nama;
    }

    public int getUmur() {
        return umur;
    }

    public void setUmur(int umur) {
        this.umur = umur;
    }

    public mahasiswa(String nrp, String nama, int umur) {
        this.nrp = nrp;
        this.nama = nama;
        this.umur = umur;
    }
    
}
